CHET-451: [core] move migration stack parameter gathering code into new package
