+++ core/src/main/kotlin/com/atlassian/migration/datacenter/core/aws/infrastructure/migrationStack/MigrationStackInputGatheringStrategy.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.atlassian.migration.datacenter.core.aws.infrastructure
+package com.atlassian.migration.datacenter.core.aws.infrastructure.migrationStack
 
 import software.amazon.awssdk.services.cloudformation.model.Stack
