+++ core/src/main/kotlin/com/atlassian/migration/datacenter/core/aws/infrastructure/migrationStack/QuickstartStandaloneMigrationStackInputGatheringStrategy.kt
@@ -14,14 +14,17 @@
  * limitations under the License.
  */
 
-package com.atlassian.migration.datacenter.core.aws.infrastructure
+package com.atlassian.migration.datacenter.core.aws.infrastructure.migrationStack
 
 import com.atlassian.migration.datacenter.core.aws.CfnApi
-import com.atlassian.migration.datacenter.core.aws.infrastructure.QuickstartDeploymentService.DATABASE_ENDPOINT_ADDRESS_STACK_OUTPUT_KEY
-import com.atlassian.migration.datacenter.core.aws.infrastructure.QuickstartDeploymentService.SECURITY_GROUP_NAME_STACK_OUTPUT_KEY
+import com.atlassian.migration.datacenter.core.aws.infrastructure.QuickstartDeploymentService
 import software.amazon.awssdk.services.cloudformation.model.Stack
 
-class QuickstartWithVPCMigrationStackInputGatheringStrategy(private val cfnApi: CfnApi) : MigrationStackInputGatheringStrategy {
+class QuickstartStandaloneMigrationStackInputGatheringStrategy(private val cfnApi: CfnApi) : MigrationStackInputGatheringStrategy {
+
+    private val dbEndpointAddressStackOutputKey = "DBEndpointAddress"
+    private val securityGroupStackOutputKey = "SGname"
+
     override fun gatherMigrationStackInputsFromApplicationStack(stack: Stack): Map<String, String> {
         val applicationStackOutputsMap = stack.outputs().associateBy({ it.outputKey() }, { it.outputValue() })
 
@@ -34,21 +37,15 @@ class QuickstartWithVPCMigrationStackInputGatheringStrategy(private val cfnApi: 
         val cfnExports = cfnApi.exports
 
         val applicationResources = cfnApi.getStackResources(stack.stackName())
-
-        val jiraStack = applicationResources["JiraDCStack"]
-        val jiraStackName = jiraStack!!.physicalResourceId()
-
-        val jiraResources = cfnApi.getStackResources(jiraStackName)
-        val efsId = jiraResources["ElasticFileSystem"]!!.physicalResourceId()
+        val efsId = applicationResources["ElasticFileSystem"]!!.physicalResourceId()
 
         return mapOf(
                 "NetworkPrivateSubnet" to cfnExports["${exportPrefix}PriNets"]!!.split(",")[0],
                 "EFSFileSystemId" to efsId,
-                "EFSSecurityGroup" to applicationStackOutputsMap[SECURITY_GROUP_NAME_STACK_OUTPUT_KEY]!!,
-                "RDSSecurityGroup" to applicationStackOutputsMap[SECURITY_GROUP_NAME_STACK_OUTPUT_KEY]!!,
-                "RDSEndpoint" to applicationStackOutputsMap[DATABASE_ENDPOINT_ADDRESS_STACK_OUTPUT_KEY]!!,
+                "EFSSecurityGroup" to applicationStackOutputsMap[securityGroupStackOutputKey]!!,
+                "RDSSecurityGroup" to applicationStackOutputsMap[securityGroupStackOutputKey]!!,
+                "RDSEndpoint" to applicationStackOutputsMap[dbEndpointAddressStackOutputKey]!!,
                 "HelperInstanceType" to "c5.large",
                 "HelperVpcId" to cfnExports["${exportPrefix}VPCID"]!!
-        )
-    }
+        )    }
 }
