+++ core/src/main/kotlin/com/atlassian/migration/datacenter/core/aws/infrastructure/migrationStack/QuickstartWithVPCMigrationStackInputGatheringStrategy.kt
@@ -14,12 +14,16 @@
  * limitations under the License.
  */
 
-package com.atlassian.migration.datacenter.core.aws.infrastructure
+package com.atlassian.migration.datacenter.core.aws.infrastructure.migrationStack
 
 import com.atlassian.migration.datacenter.core.aws.CfnApi
 import software.amazon.awssdk.services.cloudformation.model.Stack
 
-class QuickstartStandaloneMigrationStackInputGatheringStrategy(private val cfnApi: CfnApi) : MigrationStackInputGatheringStrategy {
+class QuickstartWithVPCMigrationStackInputGatheringStrategy(private val cfnApi: CfnApi) : MigrationStackInputGatheringStrategy {
+
+    private val dbEndpointAddressStackOutputKey = "DBEndpointAddress"
+    private val securityGroupStackOutputKey = "SGname"
+
     override fun gatherMigrationStackInputsFromApplicationStack(stack: Stack): Map<String, String> {
         val applicationStackOutputsMap = stack.outputs().associateBy({ it.outputKey() }, { it.outputValue() })
 
@@ -32,15 +36,21 @@ class QuickstartStandaloneMigrationStackInputGatheringStrategy(private val cfnAp
         val cfnExports = cfnApi.exports
 
         val applicationResources = cfnApi.getStackResources(stack.stackName())
-        val efsId = applicationResources["ElasticFileSystem"]!!.physicalResourceId()
+
+        val jiraStack = applicationResources["JiraDCStack"]
+        val jiraStackName = jiraStack!!.physicalResourceId()
+
+        val jiraResources = cfnApi.getStackResources(jiraStackName)
+        val efsId = jiraResources["ElasticFileSystem"]!!.physicalResourceId()
 
         return mapOf(
                 "NetworkPrivateSubnet" to cfnExports["${exportPrefix}PriNets"]!!.split(",")[0],
                 "EFSFileSystemId" to efsId,
-                "EFSSecurityGroup" to applicationStackOutputsMap[QuickstartDeploymentService.SECURITY_GROUP_NAME_STACK_OUTPUT_KEY]!!,
-                "RDSSecurityGroup" to applicationStackOutputsMap[QuickstartDeploymentService.SECURITY_GROUP_NAME_STACK_OUTPUT_KEY]!!,
-                "RDSEndpoint" to applicationStackOutputsMap[QuickstartDeploymentService.DATABASE_ENDPOINT_ADDRESS_STACK_OUTPUT_KEY]!!,
+                "EFSSecurityGroup" to applicationStackOutputsMap[securityGroupStackOutputKey]!!,
+                "RDSSecurityGroup" to applicationStackOutputsMap[securityGroupStackOutputKey]!!,
+                "RDSEndpoint" to applicationStackOutputsMap[dbEndpointAddressStackOutputKey]!!,
                 "HelperInstanceType" to "c5.large",
                 "HelperVpcId" to cfnExports["${exportPrefix}VPCID"]!!
-        )    }
+        )
+    }
 }
